Entities folder fixes
