+++ peace-erp/Data/Features/AccountTypes/Repositories/Disk/Entities/AccountType.cs
@@ -1,0 +1,8 @@
+﻿using peace_erp.Data.Database.Base;
+
+namespace peace_erp.Data.Features.AccountTypes.Repositories.Disk.Entities;
+
+public class AccountType : BaseEntity
+{
+    public string Description { get; set; }
+}
