+++ peace-erp/Data/Features/PersonLegals/Repositories/Disk/Entities/PersonLegal.cs
@@ -1,0 +1,8 @@
+﻿using peace_erp.Data.Database.Base;
+
+namespace peace_erp.Data.Features.PersonLegals.Repositories.Disk.Entities;
+
+public class PersonLegal : BaseEntity
+{
+    public string Description { get; set; }
+}
