+++ peace-erp/Data/Features/PersonTypes/Repositories/Disk/Entities/PersonType.cs
@@ -1,0 +1,8 @@
+﻿using peace_erp.Data.Database.Base;
+
+namespace peace_erp.Data.Features.PersonTypes.Repositories.Disk.Entities;
+
+public class PersonType : BaseEntity
+{
+    public string Description { get; set; }
+}
