--- peace-erp/Domain/Entities/AccountTypes.cs
@@ -1,8 +1,0 @@
-﻿using peace_erp.Domain.Entities.Base;
-
-namespace peace_erp.Domain.Entities;
-
-public class AccountTypes : BaseEntity
-{
-    public string Description { get; set; }
-}
