--- peace-erp/Domain/Entities/BilingTypes.cs
@@ -1,9 +1,0 @@
-﻿using peace_erp.Domain.Entities.Base;
-
-namespace peace_erp.Domain.Entities;
-
-public class BilingTypes : BaseEntity
-{
-    public string Description { get; set; }
-    public ShownAt ShownAt { get; set; }
-}
