--- peace-erp/Domain/Entities/PersonLegals.cs
@@ -1,8 +1,0 @@
-﻿using peace_erp.Domain.Entities.Base;
-
-namespace peace_erp.Domain.Entities;
-
-public class PersonLegals : BaseEntity
-{
-    public string Description { get; set; }
-}
