+++ peace-erp/peace-erp.csproj
@@ -22,6 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Data\Features\NovaPasta\" />
     <Folder Include="Presentation\" />
   </ItemGroup>
